+++ shoop/utils/_unitted_decimal.py
@@ -9,7 +9,7 @@ from __future__ import unicode_literals
 import decimal
 
 
-class UnitedDecimal(decimal.Decimal):
+class UnittedDecimal(decimal.Decimal):
     """
     Decimal with unit.
 
@@ -19,7 +19,7 @@ class UnitedDecimal(decimal.Decimal):
         TaxfulPrice(1) + TaxlessPrice(2)
 
     where `TaxfulPrice` and `TaxlessPrice` are subclasses of
-    `UnitedDecimal`.
+    `UnittedDecimal`.
     """
     @property
     def value(self):
@@ -31,7 +31,7 @@ class UnitedDecimal(decimal.Decimal):
         return decimal.Decimal(self)
 
     def __repr__(self):
-        decimal_repr = super(UnitedDecimal, self).__repr__()
+        decimal_repr = super(UnittedDecimal, self).__repr__()
         return decimal_repr.replace('Decimal', type(self).__name__)
 
     def unit_matches_with(self, other):
@@ -51,7 +51,7 @@ class UnitedDecimal(decimal.Decimal):
 
         :type value:
         :return: Object with same type as self and matching unit, but with given decimal value
-        :rtype: UnitedDecimal
+        :rtype: UnittedDecimal
         """
         return type(self)(value)
 
@@ -61,42 +61,42 @@ class UnitedDecimal(decimal.Decimal):
 
     def __lt__(self, other, **kwargs):
         self._check_units_match(other)
-        return super(UnitedDecimal, self).__lt__(other, **kwargs)
+        return super(UnittedDecimal, self).__lt__(other, **kwargs)
 
     def __le__(self, other, **kwargs):
         self._check_units_match(other)
-        return super(UnitedDecimal, self).__le__(other, **kwargs)
+        return super(UnittedDecimal, self).__le__(other, **kwargs)
 
     def __gt__(self, other, **kwargs):
         self._check_units_match(other)
-        return super(UnitedDecimal, self).__gt__(other, **kwargs)
+        return super(UnittedDecimal, self).__gt__(other, **kwargs)
 
     def __ge__(self, other, **kwargs):
         self._check_units_match(other)
-        return super(UnitedDecimal, self).__ge__(other, **kwargs)
+        return super(UnittedDecimal, self).__ge__(other, **kwargs)
 
     def __eq__(self, other, *args, **kwargs):
         if not self.unit_matches_with(other):
             return False
-        return super(UnitedDecimal, self).__eq__(other, **kwargs)
+        return super(UnittedDecimal, self).__eq__(other, **kwargs)
 
     def __ne__(self, other, *args, **kwargs):
         if not self.unit_matches_with(other):
             return True
-        return super(UnitedDecimal, self).__ne__(other, **kwargs)
+        return super(UnittedDecimal, self).__ne__(other, **kwargs)
 
     def __add__(self, other, **kwargs):
         self._check_units_match(other)
-        return self.new(super(UnitedDecimal, self).__add__(other, **kwargs))
+        return self.new(super(UnittedDecimal, self).__add__(other, **kwargs))
 
     def __sub__(self, other, **kwargs):
         self._check_units_match(other)
-        return self.new(super(UnitedDecimal, self).__sub__(other, **kwargs))
+        return self.new(super(UnittedDecimal, self).__sub__(other, **kwargs))
 
     def __mul__(self, other, **kwargs):
-        if isinstance(other, UnitedDecimal):
+        if isinstance(other, UnittedDecimal):
             raise TypeError('Cannot multiply %r with %r' % (self, other))
-        return self.new(super(UnitedDecimal, self).__mul__(other, **kwargs))
+        return self.new(super(UnittedDecimal, self).__mul__(other, **kwargs))
 
     def __radd__(self, other, **kwargs):
         return self.__add__(other, **kwargs)
@@ -108,52 +108,52 @@ class UnitedDecimal(decimal.Decimal):
         return self.__mul__(other, **kwargs)
 
     def __truediv__(self, other, **kwargs):
-        if isinstance(other, UnitedDecimal):
+        if isinstance(other, UnittedDecimal):
             self._check_units_match(other)
-            return super(UnitedDecimal, self).__truediv__(other, **kwargs)
+            return super(UnittedDecimal, self).__truediv__(other, **kwargs)
         else:
-            value = super(UnitedDecimal, self).__truediv__(other, **kwargs)
+            value = super(UnittedDecimal, self).__truediv__(other, **kwargs)
             return self.new(value)
 
     def __rtruediv__(self, other, **kwargs):
-        if not isinstance(other, UnitedDecimal):
+        if not isinstance(other, UnittedDecimal):
             type_name = type(self).__name__
             raise TypeError('Cannot divide non-{0} with {0}'.format(type_name))
         self._check_units_match(other)
-        return super(UnitedDecimal, self).__rtruediv__(other, **kwargs)
+        return super(UnittedDecimal, self).__rtruediv__(other, **kwargs)
 
     __div__ = __truediv__
     __rdiv__ = __rtruediv__
 
     def __floordiv__(self, other, **kwargs):
-        if not isinstance(other, UnitedDecimal):
+        if not isinstance(other, UnittedDecimal):
             type_name = type(self).__name__
             msg = 'Cannot floor-div {0} with non-{0}'.format(type_name)
             raise TypeError(msg)
         self._check_units_match(other)
-        return super(UnitedDecimal, self).__floordiv__(other, **kwargs)
+        return super(UnittedDecimal, self).__floordiv__(other, **kwargs)
 
     def __rfloordiv__(self, other, **kwargs):
-        if not isinstance(other, UnitedDecimal):
+        if not isinstance(other, UnittedDecimal):
             type_name = type(self).__name__
             msg = 'Cannot floor-div non-{0} with {0}'.format(type_name)
             raise TypeError(msg)
         self._check_units_match(other)
-        return super(UnitedDecimal, self).__rfloordiv__(other, **kwargs)
+        return super(UnittedDecimal, self).__rfloordiv__(other, **kwargs)
 
     def __mod__(self, other, **kwargs):
-        if not isinstance(other, UnitedDecimal):
+        if not isinstance(other, UnittedDecimal):
             type_name = type(self).__name__
             raise TypeError('Cannot modulo {0} with non-{0}'.format(type_name))
         self._check_units_match(other)
-        return self.new(super(UnitedDecimal, self).__mod__(other, **kwargs))
+        return self.new(super(UnittedDecimal, self).__mod__(other, **kwargs))
 
     def __divmod__(self, other, **kwargs):
-        if not isinstance(other, UnitedDecimal):
+        if not isinstance(other, UnittedDecimal):
             type_name = type(self).__name__
             raise TypeError('Cannot divmod {0} with non-{0}'.format(type_name))
         self._check_units_match(other)
-        (div, mod) = super(UnitedDecimal, self).__divmod__(other, **kwargs)
+        (div, mod) = super(UnittedDecimal, self).__divmod__(other, **kwargs)
         return (div, self.new(mod))
 
     def __pow__(self, other, **kwargs):
@@ -161,42 +161,42 @@ class UnitedDecimal(decimal.Decimal):
         raise TypeError('{} cannot be powered'.format(type_name))
 
     def __neg__(self, **kwargs):
-        return self.new(super(UnitedDecimal, self).__neg__(**kwargs))
+        return self.new(super(UnittedDecimal, self).__neg__(**kwargs))
 
     def __pos__(self, **kwargs):
-        return self.new(super(UnitedDecimal, self).__pos__(**kwargs))
+        return self.new(super(UnittedDecimal, self).__pos__(**kwargs))
 
     def __abs__(self, **kwargs):
-        return self.new(super(UnitedDecimal, self).__abs__(**kwargs))
+        return self.new(super(UnittedDecimal, self).__abs__(**kwargs))
 
     def __int__(self, **kwargs):
-        return super(UnitedDecimal, self).__int__(**kwargs)
+        return super(UnittedDecimal, self).__int__(**kwargs)
 
     def __float__(self, **kwargs):
-        return super(UnitedDecimal, self).__float__(**kwargs)
+        return super(UnittedDecimal, self).__float__(**kwargs)
 
     def __round__(self, ndigits=0, **kwargs):
-        value = super(UnitedDecimal, self).__round__(ndigits, **kwargs)
+        value = super(UnittedDecimal, self).__round__(ndigits, **kwargs)
         return self.new(value)
 
     def quantize(self, exp, *args, **kwargs):
-        value = super(UnitedDecimal, self).quantize(exp, *args, **kwargs)
+        value = super(UnittedDecimal, self).quantize(exp, *args, **kwargs)
         return self.new(value)
 
     def copy_negate(self, *args, **kwargs):
-        value = super(UnitedDecimal, self).copy_negate(*args, **kwargs)
+        value = super(UnittedDecimal, self).copy_negate(*args, **kwargs)
         return self.new(value)
 
 
 class UnitMixupError(TypeError):
     """
-    Invoked operation for UnitedDecimal and object with non-matching unit.
+    Invoked operation for UnittedDecimal and object with non-matching unit.
 
     The objects involved are stored in instance variables `obj1` and
-    `obj2`.  Former is instance of :class:`UnitedDecimal` or its
+    `obj2`.  Former is instance of :class:`UnittedDecimal` or its
     subclass and the other could be any object.
 
-    :ivar UnitedDecimal obj1: Involved object 1
+    :ivar UnittedDecimal obj1: Involved object 1
     :ivar Any obj2: Involved object 2
     """
     def __init__(self, obj1, obj2, msg='Unit mixup'):